Added new features, more error handling
Added a new "manual mode" which will stop the scraper side of the scraper poster. This allows you to log in to instagram without causing problems to the bot. Once you are done you can simply resume the bot.

Added the possibility of commenting on the post after uploading it. This feature can be probably expanded upon, by using more unique comments, and comments structure (like commenting on your comment might be cool)